Avoid SHMEM bottleneck in largedata.
We have to figureout why the receive_buffer_size is not overwritten in setup_[DW, DR]_qos in order to know if there are other QoS that are being ignored.

Still the receive_buffer_size and received_message_count_max QoS settings should be dynamically configured in PerftestTransfort.cxx according to some kind of rule instead of having it hardcoded within the code.